master

+++ helper/src/main/java/com/vsg/helper/common/model/IId.kt
@@ -1,5 +1,5 @@
 package com.vsg.helper.common.model
 
 interface IId {
-    var id: Long
+    var id: Int
 }
